Changed all int instances to double to reflect price

+++ src/main/java/entity/Recipe.java
@@ -6,17 +6,17 @@ package entity;
  */
 public class Recipe {
     private final String name;
-    private final int basePrice;
-    private int currentPrice;
+    private final double basePrice;
+    private double currentPrice;
     private int stock;
 
     /**
-     * Creates a new recipe with the given non-empty name and positive integer basePrice.
+     * Creates a new recipe with the given non-empty name and positive double basePrice.
      * @param name the name
      * @param basePrice the basePrice
      * @throws IllegalArgumentException if the name is empty or if basePrice is non-positive
      */
-    public Recipe(String name, int basePrice) {
+    public Recipe(String name, double basePrice) {
         if ("".equals(name)) {
             throw new IllegalArgumentException("Name cannot be empty");
         }
@@ -44,15 +44,15 @@ public class Recipe {
         return stock;
     }
 
-    public int getBasePrice() {
+    public double getBasePrice() {
         return basePrice;
     }
 
-    public int getCurrentPrice() {
+    public double getCurrentPrice() {
         return currentPrice;
     }
 
-    public void setCurrentPrice(int newPrice) {
+    public void setCurrentPrice(double newPrice) {
         if (newPrice <= 0) {
             throw new IllegalArgumentException("Price must be positive");
         }
